Backend/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
astunparse==1.6.3
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.12.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dm-tree==0.1.8
dnspython==2.6.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Session==0.8.0
flatbuffers==24.3.7
fonttools==4.51.0
gast==0.5.4
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==22.0.0
h5py==3.10.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.0
keras==3.0.5
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
msgspec==0.18.6
namex==0.0.7
numpy==1.26.4
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pymongo==4.6.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rich==13.7.1
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
setuptools==69.1.1
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
#tensorflow==2.17.0
#tensorflow-intel==2.17.0
termcolor==2.4.0
threadpoolctl==3.4.0
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.16.0
